requirements.txt
aioschedule==0.5.2
anyio==3.6.2
APScheduler==3.6.3
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
praw==7.6.1
prawcore==2.3.0
python-telegram-bot==20.0a6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tornado==6.1
tzdata==2022.7
tzlocal==4.2
update-checker==0.18.0
urllib3==1.26.13
websocket-client==1.4.2
